Extract sendExpiryEmail code to standalone file
