+++ controllers/apply/expiries/views/awards-for-all-expiry-email.njk
@@ -1,4 +1,4 @@
-{% extends "views/layouts/emails.njk" %}
+{% extends "../../../../views/layouts/emails.njk" %}
 
 {% block content %}
