+++ controllers/apply/expiries/views/standard-expiry-email.njk
@@ -1,4 +1,4 @@
-{% extends "views/layouts/emails.njk" %}
+{% extends "../../../../views/layouts/emails.njk" %}
 
 {% block content %}
